Remarks gathered during code review applied to tests related to audit logs.
