+++ src/integrationTest/java/org/opensearch/security/SearchOperationTest.java
@@ -96,9 +96,9 @@ import org.opensearch.index.reindex.ReindexRequest;
 import org.opensearch.repositories.RepositoryMissingException;
 import org.opensearch.rest.RestStatus;
 import org.opensearch.search.builder.SearchSourceBuilder;
-import org.opensearch.test.framework.TestSecurityConfig.AuditCompliance;
-import org.opensearch.test.framework.TestSecurityConfig.AuditConfiguration;
-import org.opensearch.test.framework.TestSecurityConfig.AuditFilters;
+import org.opensearch.test.framework.AuditCompliance;
+import org.opensearch.test.framework.AuditConfiguration;
+import org.opensearch.test.framework.AuditFilters;
 import org.opensearch.test.framework.TestSecurityConfig.Role;
 import org.opensearch.test.framework.TestSecurityConfig.User;
 import org.opensearch.test.framework.audit.AuditLogsRule;
